+++ components/Comment/Comment.module.css
@@ -1,0 +1,113 @@
+.comment_register_container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin-bottom: 20px;
+}
+
+.comment_register_container textarea {
+  text-indent: 0;
+  padding-left: 20px;
+}
+
+.comment_textarea,
+.comment_edit_textarea {
+  border: none;
+  border-radius: 10px;
+  background-color: #f3f4f6;
+  height: 104px;
+  font-family: sans-serif;
+  box-sizing: border-box;
+  resize: none;
+  padding-top: 20px;
+  text-indent: 20px;
+}
+
+.comment_edit_textarea {
+  width: 90%;
+}
+
+.button_container {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.register_button,
+.edit_button {
+  width: 74px;
+  height: 42px;
+  border: none;
+  border-radius: 7px;
+  background-color: #9ca3af;
+  color: white;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.register_button.active_button {
+  background-color: #3692ff;
+  cursor: pointer;
+}
+
+.comment_title {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 15px;
+}
+
+.edit_container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  gap: 20px;
+}
+
+.edit_container textarea {
+  width: 100%;
+}
+
+.btn_container {
+  display: flex;
+  justify-content: flex-end;
+  gap: 20px;
+}
+
+.cancel_button {
+  color: #737373;
+  cursor: pointer;
+  background: none;
+  border: none;
+}
+
+.edit_button {
+  background-color: #3692ff;
+  width: 106px;
+}
+
+.select_button {
+  cursor: pointer;
+}
+
+.detail_user_stats {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  padding-bottom: 10px;
+  border-bottom: 1px solid #e5e7eb;
+  margin-bottom: 20px;
+}
+
+.user_img {
+  width: 40px;
+  height: 40px;
+}
+
+.nickname {
+  color: #4b5563;
+}
+
+.create_at {
+  color: #9ca3af;
+  font-weight: lighter;
+}
