+++ pages/board/styles/detail.module.css
@@ -51,34 +51,15 @@
   gap: 20px;
   margin-bottom: 20px;
 }
-.comment_textarea,
-.comment_edit_textarea {
-  border: none;
-  border-radius: 10px;
-  background-color: #f3f4f6;
-  height: 104px;
-  font-family: sans-serif;
-  box-sizing: border-box;
-  resize: none;
-  padding-top: 20px;
-  text-indent: 20px;
-}
 
-.button_container {
-  display: flex;
-  justify-content: flex-end;
+.comment_register_container textarea {
+  text-indent: 0;
+  padding-left: 20px;
 }
 
 .comment_container {
   background-color: #fcfcfc;
   gap: 10px;
-}
-
-.comment_title {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 15px;
 }
 
 .comment_user_stats {
@@ -109,40 +90,6 @@
   display: flex;
   justify-content: center;
   margin-top: 50px;
-}
-
-.comment_edit_textarea {
-  width: 90%;
-}
-
-.register_button,
-.edit_button {
-  width: 74px;
-  height: 42px;
-  border: none;
-  border-radius: 7px;
-  background-color: #9ca3af;
-  color: white;
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.register_button.active_button {
-  background-color: #3692ff;
-  cursor: pointer;
-}
-
-.edit_container {
-  display: flex;
-  flex-direction: column;
-}
-.edit_container {
-  width: 100%;
-  gap: 20px;
-}
-
-.edit_container textarea {
-  width: 100%;
 }
 
 .btn_container {
